requirements.txt
# requirements.txt

# Core Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
starlette>=0.36.0
websockets>=12.0

# Data Processing & ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
tensorflow>=2.15.0
joblib>=1.3.0

# Scraping & Async
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
backoff>=2.2.0
httpx>=0.26.0

# Azure Integration
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1
loguru>=0.7.2